// src/generated/api/models/PagedResponsePostListResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 모든 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PostListResponse } from './PostListResponse';
import {
    PostListResponseFromJSON,
    PostListResponseFromJSONTyped,
    PostListResponseToJSON,
    PostListResponseToJSONTyped,
} from './PostListResponse';

/**
 * 
 * @export
 * @interface PagedResponsePostListResponse
 */
export interface PagedResponsePostListResponse {
    /**
     * 
     * @type {Array<PostListResponse>}
     * @memberof PagedResponsePostListResponse
     */
    content?: Array<PostListResponse>;
    /**
     * 
     * @type {number}
     * @memberof PagedResponsePostListResponse
     */
    page?: number;
    /**
     * 
     * @type {number}
     * @memberof PagedResponsePostListResponse
     */
    size?: number;
    /**
     * 
     * @type {number}
     * @memberof PagedResponsePostListResponse
     */
    totalElements?: number;
    /**
     * 
     * @type {number}
     * @memberof PagedResponsePostListResponse
     */
    totalPages?: number;
}

/**
 * Check if a given object implements the PagedResponsePostListResponse interface.
 */
export function instanceOfPagedResponsePostListResponse(value: object): value is PagedResponsePostListResponse {
    return true;
}

export function PagedResponsePostListResponseFromJSON(json: any): PagedResponsePostListResponse {
    return PagedResponsePostListResponseFromJSONTyped(json, false);
}

export function PagedResponsePostListResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PagedResponsePostListResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'content': json['content'] == null ? undefined : ((json['content'] as Array<any>).map(PostListResponseFromJSON)),
        'page': json['page'] == null ? undefined : json['page'],
        'size': json['size'] == null ? undefined : json['size'],
        'totalElements': json['totalElements'] == null ? undefined : json['totalElements'],
        'totalPages': json['totalPages'] == null ? undefined : json['totalPages'],
    };
}

export function PagedResponsePostListResponseToJSON(json: any): PagedResponsePostListResponse {
    return PagedResponsePostListResponseToJSONTyped(json, false);
}

export function PagedResponsePostListResponseToJSONTyped(value?: PagedResponsePostListResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'content': value['content'] == null ? undefined : ((value['content'] as Array<any>).map(PostListResponseToJSON)),
        'page': value['page'],
        'size': value['size'],
        'totalElements': value['totalElements'],
        'totalPages': value['totalPages'],
    };
}