// src/generated/api/models/PostPopularResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 모든 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0.4
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface PostPopularResponse
 */
export interface PostPopularResponse {
  /**
   *
   * @type {number}
   * @memberof PostPopularResponse
   */
  postId?: number;
  /**
   *
   * @type {string}
   * @memberof PostPopularResponse
   */
  title?: string;
  /**
   *
   * @type {string}
   * @memberof PostPopularResponse
   */
  writer?: string;
  /**
   *
   * @type {string}
   * @memberof PostPopularResponse
   */
  writerProfileImageUrl?: string;
}

/**
 * Check if a given object implements the PostPopularResponse interface.
 */
export function instanceOfPostPopularResponse(value: object): value is PostPopularResponse {
  return true;
}

export function PostPopularResponseFromJSON(json: any): PostPopularResponse {
  return PostPopularResponseFromJSONTyped(json, false);
}

export function PostPopularResponseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): PostPopularResponse {
  if (json == null) {
    return json;
  }
  return {
    postId: json["postId"] == null ? undefined : json["postId"],
    title: json["title"] == null ? undefined : json["title"],
    writer: json["writer"] == null ? undefined : json["writer"],
    writerProfileImageUrl:
      json["writerProfileImageUrl"] == null ? undefined : json["writerProfileImageUrl"],
  };
}

export function PostPopularResponseToJSON(json: any): PostPopularResponse {
  return PostPopularResponseToJSONTyped(json, false);
}

export function PostPopularResponseToJSONTyped(
  value?: PostPopularResponse | null,
  ignoreDiscriminator: boolean = false
): any {
  if (value == null) {
    return value;
  }

  return {
    postId: value["postId"],
    title: value["title"],
    writer: value["writer"],
    writerProfileImageUrl: value["writerProfileImageUrl"],
  };
}