// src/generated/api/models/PostRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 모든 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0.4
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface PostRequest
 */
export interface PostRequest {
  /**
   *
   * @type {string}
   * @memberof PostRequest
   */
  title: string;
  /**
   *
   * @type {string}
   * @memberof PostRequest
   */
  content: string;
  /**
   *
   * @type {Array<string>}
   * @memberof PostRequest
   */
  tags: Array<string>;
  /**
   *
   * @type {string}
   * @memberof PostRequest
   */
  commitMessage?: string;
}

/**
 * Check if a given object implements the PostRequest interface.
 */
export function instanceOfPostRequest(value: object): value is PostRequest {
  if (!("title" in value) || value["title"] === undefined) return false;
  if (!("content" in value) || value["content"] === undefined) return false;
  if (!("tags" in value) || value["tags"] === undefined) return false;
  return true;
}

export function PostRequestFromJSON(json: any): PostRequest {
  return PostRequestFromJSONTyped(json, false);
}

export function PostRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): PostRequest {
  if (json == null) {
    return json;
  }
  return {
    title: json["title"],
    content: json["content"],
    tags: json["tags"],
    commitMessage: json["commitMessage"] == null ? undefined : json["commitMessage"],
  };
}

export function PostRequestToJSON(json: any): PostRequest {
  return PostRequestToJSONTyped(json, false);
}

export function PostRequestToJSONTyped(
  value?: PostRequest | null,
  ignoreDiscriminator: boolean = false
): any {
  if (value == null) {
    return value;
  }

  return {
    title: value["title"],
    content: value["content"],
    tags: value["tags"],
    commitMessage: value["commitMessage"],
  };
}